Performance: changed the expression parser.
I got rid of the Then call in the returned parser. Instead we build a func in each of the Infix parsers and apply it to the first term.

This performs a bit worse on inputs that are long associative expressions, because we have to store the parsed results in a list, but much better on short inputs or those with a large number of subexpressions, because we're no longer building parsers dynamically on every call to Parse.