Fix performance bug when building big parsers.
Basically, when you're building a parser which involves a lot of calls
to ExpectedUtil.Concat, we can save a lot of copying using ImmutableList
to get sharing. Creates more garbage but runs ~10x faster in my test.

To do: when the right-hand part of Concat is large this still performs badly.
Perhaps use a rope-like structure instead?